Updated new CheckerPage
